pyproject.toml
[project]
name = "danbo-nov"
version = "0.1.0"
description = "AI Image Generation Prompt Optimization Web System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "optuna>=3.6.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.22",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
